refactor(lisk): move `getAccount()` to `lisk-utils`
