+++ src/components/ExportKeysForm.vue
@@ -103,7 +103,7 @@ import { validateMnemonic } from 'bip39'
 import copyToClipboard from 'copy-to-clipboard'
 import { getAccountFromPassphrase as getEthAccount } from '@/lib/eth-utils'
 import { getAccount as getBtcAccount } from '@/lib/bitcoin/btc-base-api'
-import { getAccount as getLskAccount } from '@/lib/lisk/lisk-api'
+import { getAccount as getLskAccount } from '@/lib/lisk/lisk-utils'
 import { Cryptos, CryptosInfo } from '@/lib/constants'
 import QrcodeCapture from '@/components/QrcodeCapture'
 import QrcodeScannerDialog from '@/components/QrcodeScannerDialog'
