+++ src/lib/lisk/lisk-constants.ts
@@ -1,0 +1,6 @@
+export const LiskHashSettings = {
+  SALT: 'adm',
+  ITERATIONS: 2048,
+  KEYLEN: 32,
+  DIGEST: 'sha256'
+} as const
